chore: refactor structure of destination related form file
