+++ src/components/forms/connector/destination/CreateDestinationForm/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CreateDestinationForm";
+export * from "./CreateDestinationForm";
