+++ src/components/forms/connector/destination/index.ts
@@ -1,0 +1,14 @@
+import CreateDestinationForm from "./CreateDestinationForm";
+import type { CreateDestinationFormValues } from "./CreateDestinationForm";
+import ConfigureDestinationForm from "./ConfigureDestinationForm";
+import type {
+  ConfigureDestinationFormProps,
+  ConfigureDestinationFormValue,
+} from "./ConfigureDestinationForm";
+
+export { CreateDestinationForm, ConfigureDestinationForm };
+export type {
+  CreateDestinationFormValues,
+  ConfigureDestinationFormProps,
+  ConfigureDestinationFormValue,
+};
