+++ src/components/forms/connector/index.ts
@@ -1,1 +1,2 @@
 export * from "./source";
+export * from "./destination";
