+++ src/components/forms/index.ts
@@ -1,12 +1,12 @@
 import CreatePipelineForm from "./CreatePipelineForm";
 import type { StepNumberState } from "./CreatePipelineForm/CreatePipelineForm";
 
-import CreateDestinationForm from "./CreateDestinationForm";
+import CreateDestinationForm from "./connector/destination/CreateDestinationForm";
 import CreateSourceForm from "./connector/source/CreateSourceForm";
 import OnboardingForm from "./OnboardingForm";
 import CreateModelForm from "./CreateModelForm";
 import { ConfigureSourceForm } from "./connector";
-import ConfigureDestinationForm from "./ConfigureDestinationForm";
+import ConfigureDestinationForm from "./connector/destination/ConfigureDestinationForm";
 import ConfigureModelForm from "./ConfigureModelForm";
 import ConfigureModelInstanceForm from "./ConfigureModelInstanceForm";
 import type { ConfigureModelInstanceFormProps } from "./ConfigureModelInstanceForm";
